Spelling

+++ app/controllers/facilities_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class FacilitiesController < ApplicationController
-  before_action :set_facility, only: %i[show edit update destroy]
+  before_action :set_facility, only: %i[show edit update destroy reassign reassign_to]
   before_action :set_sector_collection, only: %i[new edit update create]
 
   # GET /facilities
@@ -108,7 +108,7 @@ class FacilitiesController < ApplicationController
 
     respond_to do |format|
       if @facility.save
-        format.html { redirect_to @facility, notice: 'Facility updated.' }
+        format.html { redirect_to @return_path, notice: 'Facility updated.' }
         format.json { render :show, status: :ok, location: @facility }
       else
         format.html { render :edit }
@@ -120,11 +120,41 @@ class FacilitiesController < ApplicationController
   # DELETE /facilities/1
   # DELETE /facilities/1.json
   def destroy
+    redirect_to reassign_facility_path(@facility) and return if @facility.reports.any? || @facility.plans.any?
+
+    # break the reassign / reassign_to loop
+    @return_path = facilities_path if @return_path.include? 'reassign'
+
     authorize @facility.destroy
     respond_to do |format|
-      format.html { redirect_to facilities_url, notice: 'Facility destroyed.' }
+      format.html { redirect_to @return_path, notice: 'Facility destroyed.' }
       format.json { head :no_content }
     end
+  end
+
+  def reassign
+    authorize @facility
+
+    @reports = @facility.reports
+    @plans = @facility.plans
+
+    if @reports.any? || @plans.any?
+      @facilities = @facility.similar_by_name.select(:name, :id, :hierarchy)
+      render :reassign
+    else
+      render :can_be_deleted
+    end
+  end
+
+  def reassign_to
+    authorize @facility
+
+    @to_facility = Facility.find(params[:to])
+
+    @facility.reports.update_all(reportable_id: @to_facility.id) if @facility.reports.any?
+    @facility.plans.update_all(planable_id: @to_facility.id) if @facility.plans.any?
+
+    redirect_to reassign_facility_path(@facility)
   end
 
   private
